requirements.txt
transformers
gradio
Pillow
torch